requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2==2.9.9
drf-standardized-errors==0.14.1
twilio==9.3.2
sendgrid==6.11.0
